fix(automation): Move metadata services to automation
Some of the request metadata services belong in automation because it
maps request metadata to events, which is the automation feature. This
was setup under the metadata flags by mistake.
