+++ CSharp/.env
@@ -9,3 +9,5 @@ $(upper_appname)_WEBAPI__HTTPS_PORT=5001
 $(upper_appname)_DB__SAPASSWORD=dfGhaj76Kjdh82983
 $(upper_appname)_DB__PORT=1433
 $(upper_appname)_CONNECTIONSTRINGS__MAIN=Server=db;User Id=sa;Password=dfGhaj76Kjdh82983;Initial Catalog=BusinessApp
+$(upper_appname)_CONNECTIONSTRINGS__TEST=Server=db;User Id=sa;Password=dfGhaj76Kjdh82983;Initial Catalog=BusinessApp_IntegrationTest
+$(upper_appname)_TEST_CONNECTIONSTRINGS__MAIN=Server=db;User Id=sa;Password=dfGhaj76Kjdh82983;Initial Catalog=BusinessApp_FunctionalTest
